% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapviews.R
\name{camtrap_operation_mapview}
\alias{camtrap_operation_mapview}
\title{View the camera trap}
\usage{
camtrap_operation_mapview(camtrap_operation_table)
}
\arguments{
\item{camtrap_operation_table}{camera trap operation table as found in the camtrap schema}
}
\value{
mapview
}
\description{
View the camera trap
}
