% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-db.R
\name{check_unique_project}
\alias{check_unique_project}
\title{check whether project is unique or already on database}
\usage{
check_unique_project(ProjectShortName, con)
}
\arguments{
\item{ProjectShortName}{A list of pointblank agents (directly from camera_trap_dq())}

\item{con}{database connection}
}
\value{
message
}
\description{
check whether project is unique or already on database
}
