% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise-coords.R
\name{convert_to_latlong}
\alias{convert_to_latlong}
\title{Convert zone 54/55 easting/northings into lat/longs}
\usage{
convert_to_latlong(data)
}
\arguments{
\item{data}{data.frame with three columns (Easting, Northing, Zone)}
}
\value{
data but with Easting, Northing and Zone replaced with 'Longitude' and 'Latitude'
}
\description{
Convert zone 54/55 easting/northings into lat/longs
}
