% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data Dictionary}
\format{
A data frame with 131 rows and 7 variables
}
\usage{
data(data_dictionary)
}
\description{
A data dictionary of the database
}
\details{
\itemize{
\item schema. Schema name
\item table_name. Name of the table
\item table_description. Description of the purpose of the table
\item table_type. Level of table processing (raw, curated or processed)
\item column_name. Name of the column/field
\item column_class. Class of the column as per R conventions
\item column_description. Description of the column purpose
\item darwin_standard_core. Darwin standard core matching field
\item derived_from. Where the data is derived from (e.g. user or camtrapR)
}
}
\keyword{datasets}
