% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-upload.R
\name{dataUploadpUI}
\alias{dataUploadpUI}
\alias{dataUploadServer}
\title{Data Upload Shiny Module}
\usage{
dataUploadpUI(id, label = "dataUpload")

dataUploadServer(id)
}
\arguments{
\item{id}{shiny samespace id}

\item{label}{label of UI}
}
\value{
shiny module
}
\description{
Data Upload Shiny Module
}
