% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-quality-transects.R
\name{filter_records_outside_transect_area}
\alias{filter_records_outside_transect_area}
\title{Filter records outside transect area}
\usage{
filter_records_outside_transect_area(records, transects, endcap = "FLAT")
}
\arguments{
\item{records}{data.frame of the records with AnimalLongitude and AnimalLatitude columns}

\item{transects}{sf data.frame of the transects}

\item{endcap}{style of the endcap buffering, see \link[sf]{st_buffer}, default is 'FLAT' meaning the searchable area does not extend past the end of the transect}
}
\value{
data.frame
}
\description{
Filter the records so that no animals fall outside of the searchable transect area
}
