% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proofsafe_format.R
\name{koala_proofsafe_format}
\alias{koala_proofsafe_format}
\alias{gg_proofsafe_format}
\title{Format proofsafe data for database}
\usage{
koala_proofsafe_format(
  proofsafe,
  gps_transects,
  sp_filter = "Koala",
  Iteration,
  SurveyMethod = c("Diurnal double-observer distance-sampling",
    "Nocturnal double-observer distance-sampling",
    "Diurnal single-observer distance-sampling",
    "Nocturnal single-observer distance-sampling"),
  MaxTruncationDistance
)

gg_proofsafe_format(
  proofsafe,
  gps_transects,
  Iteration,
  SurveyMethod = c("Diurnal double-observer distance-sampling",
    "Nocturnal double-observer distance-sampling",
    "Diurnal single-observer distance-sampling",
    "Nocturnal single-observer distance-sampling"),
  MaxTruncationDistance
)
}
\arguments{
\item{proofsafe}{data.frame of data directly from proofsafe}

\item{gps_transects}{line transects, formatted as an sf object with columns for SiteID and Transect}

\item{sp_filter}{species filter}

\item{Iteration}{integer of the season or repeat visit number of the survey (e.g. 1 for the first time a sites been surveyed or 2 if it has been surveyed once before)}

\item{SurveyMethod}{character type of method used to detect animals}

\item{MaxTruncationDistance}{The maximum distance in metres that observations were made out to (e.g. 100)}
}
\value{
list of data.frame for records and sf for transects
}
\description{
These functions format the datasets from proofsafe alongside gps data of the transect lines for koala or greater
glider (gg) surveys
}
