% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-db.R
\name{prepare_camtrap_upload}
\alias{prepare_camtrap_upload}
\title{Prepare camera trap data for upload to the database}
\usage{
prepare_camtrap_upload(agent_list)
}
\arguments{
\item{agent_list}{A list of pointblank agents (directly from camera_trap_dq())}
}
\value{
A list of data frames
}
\description{
This function takes a list of agents and prepares the data for upload to the database.
}
\examples{
\dontrun{
prepare_upload(agent_list)
}
}
