% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-transects.R
\name{prepare_transect_upload}
\alias{prepare_transect_upload}
\title{Prepare transect data for upload to the database}
\usage{
prepare_transect_upload(agent_list)
}
\arguments{
\item{agent_list}{A list of pointblank agents (directly from transect_dq())}
}
\value{
A list of data frames
}
\description{
This function takes a list of agents and prepares the data for upload to the database.
}
\examples{
\dontrun{
prepare_transect_upload(agent_list)
}
}
