% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-map.R
\name{projectMapUI}
\alias{projectMapUI}
\alias{projectMapServer}
\title{Shiny Map UI for Projects}
\usage{
projectMapUI(
  id,
  label = "projectMap",
  custom_css_path = system.file("app/styles.css", package = "weda"),
  custom_js_path = system.file("app/gomap.js", package = "weda"),
  colour_vars
)

projectMapServer(id, project_locations, con)
}
\arguments{
\item{id}{module id}

\item{label}{module label}

\item{custom_css_path}{custom css path for map}

\item{custom_js_path}{custom javascript path for module}

\item{colour_vars}{variables to colour by ("ProjectName" and species common names)}

\item{project_locations}{data.frame of survey locations}

\item{con}{database connection}
}
\value{
shiny module

shiny module
}
\description{
Shiny module creating an interactive project map
}
\section{Functions}{
\itemize{
\item \code{projectMapServer()}: 

}}
