% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proofsafe_to_operation.R
\name{proofsafe_to_operation}
\alias{proofsafe_to_operation}
\title{Proofsafe to operation}
\usage{
proofsafe_to_operation(
  data,
  sensitivity,
  delay,
  photos_per_trigger,
  quiet_period
)
}
\arguments{
\item{data}{data.frame/tbl from reading in the proofsafe results csv}

\item{sensitivity}{Sensitivity of the camera (e.g. "High", or "Very High")}

\item{delay}{Settings of the camera delay (e.g. Rapidfire)}

\item{photos_per_trigger}{Settings of how many photos per}

\item{quiet_period}{How long the camera is delayed after each trigger (e.g. usually 0)}
}
\value{
data.frame
}
\description{
Take the csv results file from proofsafe and format it into the operations table format.
}
\examples{
\dontrun{
test_op <- proofsafe_to_operation(data = data,
  sensitivity = "Very High",
.  delay = "Rapidfire",
.  photos_per_trigger = 5, quiet_period = 0)
}

}
