% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proofsafe_to_operation.R
\name{proofsafeDownloadUI}
\alias{proofsafeDownloadUI}
\alias{proofsafeDownloadServer}
\title{proofsafe formatter module}
\usage{
proofsafeDownloadUI(id)

proofsafeDownloadServer(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
shiny UI
}
\description{
proofsafe formatter module
}
