% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vba-check.R
\name{standardise_species_names}
\alias{standardise_species_names}
\title{This function standardises species names in a dataframe}
\usage{
standardise_species_names(
  recordTable,
  format = c("scientific", "common"),
  speciesCol = "Species"
)
}
\arguments{
\item{recordTable}{A dataframe containing species names}

\item{format}{The format of the species names in the dataframe}

\item{speciesCol}{The column name of the species names in the dataframe}
}
\value{
A dataframe with standardised species names
}
\description{
This function standardises species names in a dataframe
}
\examples{
\dontrun{
standardise_species_names(recordTable = recordTable,
                          format = "scientific",
                          speciesCol = "Species")
                          }
}
