% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-quality-transects.R
\name{transect_dq}
\alias{transect_dq}
\title{Transect Data Quality Checks}
\usage{
transect_dq(records, transects, project_information, progress = interactive())
}
\arguments{
\item{records}{this is the dataframe that contains the records of animals on transects}

\item{transects}{this is the dataframe that contains the information about the transect location and time it was surveyed}

\item{project_information}{this is the dataframe that contains the information about the project}

\item{progress}{show pointblank's step-by-step progress log (default: in interactive sessions only)}
}
\value{
list of pointblank objects
}
\description{
Assesses the data quality of transect records, transects and project information.
Automatically checks whether columns are present, converts them to the appropriate class and
runs a pointblank check on the data
}
