% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect-shiny-upload.R
\name{transectdataUploadpUI}
\alias{transectdataUploadpUI}
\alias{transectdataUploadServer}
\title{Data Upload Shiny Module}
\usage{
transectdataUploadpUI(id, label = "transectdataUpload")

transectdataUploadServer(id, con)
}
\arguments{
\item{id}{shiny samespace id}

\item{label}{label of UI}

\item{con}{database connection}
}
\value{
shiny module
}
\description{
Data Upload Shiny Module
}
