% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-transect-map.R
\name{transectMapUI}
\alias{transectMapUI}
\alias{transectMapServer}
\title{Shiny Map UI for Projects}
\usage{
transectMapUI(
  id,
  label = "transectMap",
  custom_css_path = system.file("app/styles.css", package = "weda"),
  custom_js_path = system.file("app/gomap.js", package = "weda"),
  colour_vars
)

transectMapServer(id, project_locations, con)
}
\arguments{
\item{id}{module id}

\item{label}{module label}

\item{custom_css_path}{custom css path for map}

\item{custom_js_path}{custom javascript path for module}

\item{colour_vars}{variables to colour by}

\item{project_locations}{data.frame of survey locations}

\item{con}{database connection}
}
\value{
shiny module

shiny module
}
\description{
Shiny module creating an interactive transect map
}
\section{Functions}{
\itemize{
\item \code{transectMapServer()}: 

}}
