% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-transects.R
\name{upload_transect_data}
\alias{upload_transect_data}
\title{Upload camera trap data to database}
\usage{
upload_transect_data(
  con,
  data_list,
  uploadername,
  tables_to_upload = c("raw_transect_records", "raw_transects",
    "raw_project_information"),
  schema = "transect"
)
}
\arguments{
\item{con}{postgresql connection to ari-dev-weda-psql-01}

\item{data_list}{list of transect records, transects and project information (output from prepare_transect_upload())}

\item{uploadername}{name of person uploading data}

\item{tables_to_upload}{vector (characters) of tables to upload. Default is all of them}

\item{schema}{schema to upload data to (options are transect or transect_dev)}
}
\description{
Upload camera trap data to database
}
\examples{
\dontrun{
con <- weda_connect(password = keyring::key_get(service = "ari-dev-weda-psql-01",
username = "psql_user"))
upload_transect_data(con = con, data_list = data_list, uploadername = "Justin Cally")
}
}
