% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vba-format.R
\name{vba_format}
\alias{vba_format}
\title{VBA Upload format for camera trap data}
\usage{
vba_format(con, ProjectShortName = "all", return_data = FALSE)
}
\arguments{
\item{con}{database connection}

\item{ProjectShortName}{ProjectShortName field}

\item{return_data}{logical flag to return data (TRUE) or sql query (default is FALSE)}
}
\value{
sql or data.frame
}
\description{
A presence-absence view of camera trap surveys using the VBA upload template
}
\examples{
 \dontrun{
con <- weda_connect(password = keyring::key_get(service = "ari-dev-weda-psql-01",
username = "psql_user"))
DBI::dbExecute(conn = con,
               paste(SQL("CREATE VIEW camtrap.processed_vba_format AS"),
               vba_format(con = con, return_data = FALSE)))
}
}
