% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vba_name_conversions}
\alias{vba_name_conversions}
\title{Name conversions between scientific and common names}
\format{
A data frame with 38,587 rows and 2 variables
}
\usage{
data(vba_name_conversions)
}
\description{
A dataset containing the scientific names and common names of taxa in the VBA
}
\details{
\itemize{
\item taxon_id. The VBA taxon ID
\item scientific_name. The scientific name in the VBA
\item common_name. The common name in the VBA
}
}
\keyword{datasets}
