% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise-transects.R
\name{visualise_records}
\alias{visualise_records}
\title{Visualise transects, records and sight-lines}
\usage{
visualise_records(records, transects, endcap = "FLAT")
}
\arguments{
\item{records}{data.frame of the records with AnimalLongitude and AnimalLatitude columns}

\item{transects}{sf of the transects searched}

\item{endcap}{style of the endcap buffering, see \link[sf]{st_buffer}, default is 'FLAT' meaning the searchable area does not extend past the end of the transect}
}
\value{
mapview widget
}
\description{
View an interactive map of the transects and records as well as the lines from where the observer sighted them and the distance between projected animal locations
}
