% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{weda_connect}
\alias{weda_connect}
\title{Database Connection}
\usage{
weda_connect(username = "psql_user", password, ...)
}
\arguments{
\item{username}{postgres user name (default is psql_user)}

\item{password}{postgres password (keyring use suggested)}

\item{...}{additional arguments passed to \code{RPostgreSQL::dbConnect()}}
}
\value{
PostgreSQL connection
}
\description{
function to easily connect to the weda database
}
\examples{
\dontrun{

weda_connect(password = keyring::key_get(service = "ari-dev-weda-psql-01", username = "psql_user"))

}
}
