% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-quality.R
\name{camera_trap_dq}
\alias{camera_trap_dq}
\title{Camera Trap Data Quality Checks}
\usage{
camera_trap_dq(camtrap_records, camtrap_operation, project_information)
}
\arguments{
\item{camtrap_records}{this is the dataframe that contains the camera trap records (recordTable from camtrapR)}

\item{camtrap_operation}{this is the dataframe that contains the information about the camera trap operation}

\item{project_information}{this is the dataframe that contains the information about the project}
}
\value{
list of pointblank objects
}
\description{
Assesses the data quality of camera trap records, operations and project information.
Automatically checks whether columns are present, converts them to the appropriate class and
runs a pointblank check on the data
}
