% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation-from-records.R
\name{camtrap_operation_from_records}
\alias{camtrap_operation_from_records}
\title{Operation from records}
\usage{
camtrap_operation_from_records(records, get_cam_model = TRUE)
}
\arguments{
\item{records}{camera trap records from camtrapR  with Species, SiteID and SubStation}

\item{get_cam_model}{logical flag (default is TRUE) of whether you want to get cam model and serial number (drive with photos needs to be attached)}
}
\value{
data.frame
}
\description{
Operation from records
}
\examples{
library(weda)
raw_camtrap_records <- camtrapR::recordTable(inDir  = system.file("dummydata/images",
package = "weda"),
                                             IDfrom = "metadata",
                                             cameraID = "directory",
                                             stationCol = "SiteID",
                                             camerasIndependent = TRUE,
                                             timeZone = Sys.timezone(location = TRUE),
                                             metadataSpeciesTag = "Species",
                                             removeDuplicateRecords = FALSE,
                                             returnFileNamesMissingTags = TRUE) \%>\%
dplyr::rename(SubStation = Camera) \%>\%
dplyr::mutate(SubStation = dplyr::case_when(SiteID == SubStation ~ NA_character_,
                                            TRUE ~ SubStation),
              Iteration = 1L)

 op_table <- camtrap_operation_from_records(raw_camtrap_records)

}
