% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation-from-records.R
\name{camtrap_operation_from_records}
\alias{camtrap_operation_from_records}
\title{Operation from records}
\usage{
camtrap_operation_from_records(records)
}
\arguments{
\item{records}{camera trap records from camtrapR}
}
\value{
data.frame
}
\description{
Operation from records
}
