% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proofsafe_format.R
\name{koala_proofsafe_format}
\alias{koala_proofsafe_format}
\alias{gg_proofsafe_format}
\title{Format proofsafe data for database}
\usage{
koala_proofsafe_format(proofsafe, gps_transects, sp_filter = "Koala")

gg_proofsafe_format(proofsafe, gps_transects)
}
\arguments{
\item{proofsafe}{data.frame of data directly from proofsafe}

\item{gps_transects}{line transects, formatted as an sf object with columns for SiteID and Transect}

\item{sp_filter}{species filter}
}
\value{
list of data.frame for records and sf for transects
}
\description{
These functions format the datasets from proofsafe alongside gps data of the transect lines for koala or greater
glider (gg) surveys
}
