% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-db.R
\name{upload_camtrap_data}
\alias{upload_camtrap_data}
\title{Upload camera trap data to database}
\usage{
upload_camtrap_data(
  con,
  data_list,
  uploadername,
  tables_to_upload = c("raw_camtrap_records", "raw_camtrap_operation",
    "raw_project_information"),
  schema = "camtrap"
)
}
\arguments{
\item{con}{postgresql connection to ari-dev-weda-psql-01}

\item{data_list}{list of camera trap records, operations and project information (output from prepare_camtrap_upload())}

\item{uploadername}{name of person uploading data}

\item{tables_to_upload}{vector (characters) of tables to upload. Default is all of them}

\item{schema}{schema to upload data to (options are camtrap or camtrap_dev)}
}
\description{
Upload camera trap data to database
}
\examples{
\dontrun{
con <- weda_connect(password = keyring::key_get(service = "ari-dev-weda-psql-01", username = "psql_user"))
upload_camtrap_data(con = con, data_list = data_list, uploadername = "Justin Cally")
}
}
