% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-db.R
\name{upload_camtrap_data}
\alias{upload_camtrap_data}
\title{Upload camera trap data to database}
\usage{
upload_camtrap_data(con, data_list, uploadername)
}
\arguments{
\item{con}{postgresql connection to ari-dev-weda-psql-01}

\item{data_list}{list of camera trap records, operations and project information (output from prepare_camtrap_upload())}

\item{uploadername}{name of person uploading data}
}
\description{
Upload camera trap data to database
}
\examples{
\dontrun{
upload_camtrap_data(con = con_odbc, data_list = data_list, uploadername = "Justin Cally")
}
}
